Create futuredsp crate
