+++ Cargo.toml
@@ -14,6 +14,7 @@ categories = ["asynchronous", "concurrency", "hardware-support", "science", "was
 [workspace]
 members = [
     ".",
+    "futuredsp",
     "pmt",
     "frontend",
     "examples/android",
@@ -78,6 +79,7 @@ config = "0.11.0"
 dirs = "4.0"
 futures = "0.3.18"
 futures-lite = "1.10.0"
+futuredsp = { path = "futuredsp", version = "0.0.1" }
 futuresdr-pmt = { path = "pmt", version = "0.0.2" }
 log = { version = "0.4", features = ["std", "max_level_debug", "release_max_level_off"] }
 lttng-ust = { version = "0.1.0", optional = true}
