+++ futuredsp/src/lib.rs
@@ -1,0 +1,3 @@
+#![cfg_attr(not(RUSTC_IS_STABLE), feature(core_intrinsics))]
+
+pub mod fir;
